Rename persistantStorage and apiClient to storageService and backendService
