+++ browser-extension/src/components/ProfilePage.tsx
@@ -5,7 +5,7 @@ import {
   Typography,
 } from "@material-ui/core";
 import { FC, useContext, useEffect, useState } from "react";
-import apiClient from "../services/api-client";
+import backendService from "../services/backend-service";
 import { Producer } from "../types";
 import AppContext from "./AppContext";
 
@@ -17,23 +17,25 @@ const ProfilePage: FC = () => {
 
   const loadProfile = async () => {
     try {
-      const producer: Producer = await apiClient.getProducer();
+      const producer: Producer = await backendService.getProducer();
       setProfile(producer);
       setUpdateProfile(producer);
     } catch (error) {
-      apiClient.handleError(error, setAuthState);
+      backendService.handleError(error, setAuthState);
     }
   };
 
   const updateProducer = async () => {
     try {
       if (updateProfile) {
-        const updatedProfile = await apiClient.updateProducer(updateProfile);
+        const updatedProfile = await backendService.updateProducer(
+          updateProfile
+        );
         setProfile(updatedProfile);
         alert("Saved");
       }
     } catch (error) {
-      apiClient.handleError(error, setAuthState);
+      backendService.handleError(error, setAuthState);
     }
   };
